Initial project skeleton
Use APL 2.0 for the license to mirror LazyVim as if/when I take
code/inspiration from there, I'm happy to use the same license.

Add a short README talking about what I'm doing here.

Set up folke/neoconf.nvim for this project and a stylua config for
formatting.

Last, but not least, add lazy.nvim as a core plugin.