Add the hooks to do more set up on start
This is largely copied from LazyVim and relies on lazy.nvim calling the
main module's setup function when loading a plugin. We now delegate that
to the `config` module and call the new setup function from there.

Since we don't have autocmds or keymaps yet, we comment those out for
now.

Commit d33be73b8d0a1f6b1dbabfc30bbac02d88d177b9 from LazyVim/LazyVim
used for reference and tweaked slightly.