feat(treesitter): add more laziness to how we load
First up, we always use `main` for Treesitter. Second, we use the new
`LazyFile` event to load when operating on a new buffer. Third, we add a
bunch of commands to trigger loading.

Last and certainly not least, we ensure that the `vim` and `vimdoc`
parsers are installed. Other parsers need to be added as part of their
language specifications.